+++ service/fs_service.go
@@ -1,4 +1,4 @@
-package spec
+package service
 
 import (
 	"os"
